Refactor imports from index
